# Optimize termination grace period