Extends the `PodEvictionResult` to include the amount of time that was allowed for the eviction. This can then be used inside of the `DrainNodeWithPods` to calculate the largest amount of time allotted and only wait that amount of time for all the pods to be gone, instead of always waiting for `--max-graceful-termination-sec`.

For example, if the autoscaler is running with `--max-graceful-termination-sec=3600`, and the pods on the node have the following termination grace periods:
* pod A: 60s
* pod B: unset
* pod C: 20s

Pod B will receive the default termination grace period of 30s. The autoscaler will wait 60s (plus `podEvictionHeadroom`) for the pods to be all gone, because pod A has the longest termination grace period of 60s.


In another example, say the pods have the following termination grace periods: 
* pod A: 7200s
* pod B: unset
* pod C: 20s

This time, pod A's grace period exceeds the max of 3600, so the eviction created by `evictPod` will cap the eviction time at 3600s. Again, pod B gets the default grace period of 30s. Since 3600 > 30 > 20, the drain will wait 3600 (plus `podEvictionHeadroom`) for the pods to be all gone.